security(cli): fix path hijacking and address PR feedback
- Resolve git executable using absolute paths only from PATH
- Filter out non-absolute entries from PATH
- Fix reservoir momentum bug by reverting to partial update loop
- Implement traversal security limits (depth and results)
- Remove stale hyperdim_simd module reference
- Compact and update progress/LEARNINGS.md
- Fix formatting to pass CI lint checks
