--- memory_retention_curve.csv
@@ -1,2 +1,0 @@
-step,beta0.0,beta0.1,beta0.2,beta0.3
-1,1,1,1,1
