+++ src/encoder.rs
@@ -327,92 +327,3 @@ impl TextEncoder {
         HVec10240::bundle(&ngram_vectors).unwrap_or_else(|_| HVec10240::zero())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_split_on_separators_basic() {
-        assert_eq!(
-            TextEncoder::split_on_separators("my_func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my-func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my.func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my/func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my::func"),
-            vec!["my", "func"]
-        );
-    }
-
-    #[test]
-    fn test_split_on_separators_consecutive() {
-        assert_eq!(
-            TextEncoder::split_on_separators("my__func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my--func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my..func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my//func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my::::func"),
-            vec!["my", "func"]
-        );
-        assert_eq!(
-            TextEncoder::split_on_separators("my._-func"),
-            vec!["my", "func"]
-        );
-    }
-
-    #[test]
-    fn test_split_on_separators_edge_cases() {
-        assert_eq!(TextEncoder::split_on_separators(""), Vec::<String>::new());
-        assert_eq!(TextEncoder::split_on_separators("_"), Vec::<String>::new());
-        assert_eq!(TextEncoder::split_on_separators("__"), Vec::<String>::new());
-        assert_eq!(TextEncoder::split_on_separators("::"), Vec::<String>::new());
-        assert_eq!(
-            TextEncoder::split_on_separators(":::"),
-            vec![":"] // :: splits, remaining : is part of the word
-        );
-        assert_eq!(TextEncoder::split_on_separators("word_"), vec!["word"]);
-        assert_eq!(TextEncoder::split_on_separators("_word"), vec!["word"]);
-        assert_eq!(TextEncoder::split_on_separators(":word"), vec![":word"]);
-        assert_eq!(TextEncoder::split_on_separators("word:"), vec!["word:"]);
-    }
-
-    #[test]
-    fn test_tokenize_code_interaction() {
-        assert_eq!(
-            TextEncoder::tokenize_code("hello_world std::vector"),
-            vec!["hello", "world", "std", "vector"]
-        );
-        assert_eq!(
-            TextEncoder::tokenize_code("  leading whitespace  "),
-            vec!["leading", "whitespace"]
-        );
-        assert_eq!(
-            TextEncoder::tokenize_code("path/to/file.rs"),
-            vec!["path", "to", "file", "rs"]
-        );
-    }
-}
