+++ src/lib.rs
@@ -78,7 +78,6 @@ mod persistence_versions;
 pub mod persistence_wasm;
 pub mod reservoir;
 mod reservoir_inertial; // ADR-0064
-mod reservoir_metrics;
 pub mod retrieval;
 pub mod semantic_bridge;
 pub mod singularity;
